+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 apidoc extension Change Log
 -----------------------
 
 - Bug #149: Fixed crash on wrongly formatted API links (cebe, santosh-1265)
+- Bug #160: Fixed parsing of '{@inheritdoc}' tag (klimov-paul)
 - Bug: Usage of deprecated `yii\base\Object` changed to `yii\base\BaseObject` allowing compatibility with PHP 7.2 (klimov-paul)
 - Enh #38: Fixed display of default values given as octal or hex notation (hiqsol)
 - Enh: Display TOC only if there is more than one headline (cebe)